requirements.txt
python-dotenv==1.0.1
tomli==2.0.1
requests==2.32.3
pydantic==2.9.2
